Codebase/requirements.txt
# ============================================================
# CV Sorting using LLMs – Python Dependencies
# ============================================================
# Install with:  pip install -r requirements.txt
# Python version: 3.10+
# ============================================================

# ----------------------------------------------------------
# LLM / LangChain orchestration
# ----------------------------------------------------------
langchain>=0.2.0
langchain-community>=0.2.0
langchain-core>=0.2.0

# ----------------------------------------------------------
# Ollama Python client (local LLM)
# ----------------------------------------------------------
ollama>=0.1.8

# ----------------------------------------------------------
# OpenAI (optional – only if using --provider openai)
# ----------------------------------------------------------
openai>=1.30.0

# ----------------------------------------------------------
# Document parsing
# ----------------------------------------------------------
pymupdf>=1.24.0          # PDF extraction (fitz)
python-docx>=1.1.0       # DOCX extraction

# ----------------------------------------------------------
# HTTP / networking
# ----------------------------------------------------------
requests>=2.31.0

# ----------------------------------------------------------
# Data utilities
# ----------------------------------------------------------
pydantic>=2.0.0

# ----------------------------------------------------------
# Charting / visualisation (optional – only needed for --charts / --report)
# ----------------------------------------------------------
matplotlib>=3.8.0        # PNG chart generation (Agg backend – no GUI required)
numpy>=1.26.0            # Array operations for grouped bar positioning
reportlab>=4.1.0         # PDF report generation (Report/Report.pdf)

# ----------------------------------------------------------
# Standard library back-ports / helpers
# ----------------------------------------------------------
# (No extra deps needed for dataclasses, json, re, pathlib)
